Forbid line ending before `->` in function signatures
This makes parsing them simpler, especially with the view of reusing
the parsing code.

To be clear, a line ending is still allowed _after_ `->` but _before_
the return type, since a return type will unambigously follow, but the
end of the parameter list may be followed (ignoring line endings) by
either `{`, `->` or nothing (for function declarations and the
upcoming function types).