Change disgusting abuse of `if`-`else` to using `==`.
We had to use `if fp {} else { ... }` to invert the condition, since
crvic doesn't have `not`. Now that we have comparisons, however, we
can replace it with a much more readble `if fp == null { ... }`. We
still don't have `not`, but, tbh, an explicit comparison to `null` is
probably clearer anyway.