Add (Victoria) string literals
The way C-style string literals are handled has also changed. Instead
of having a separate AST node for C strings, we create a string
node (now representing a Victoria string) and then wrap it in a
to-conversion node. The code generator has a special case for
string -> c_string, whereby it gets the `.start` field of the
underlying `VIC_STRING` data type (we assume it is zero-terminated,
which **is** the case for string literals).