Add `and` and `or`
In rVic, we restrict these to work only on booleans. The reason being
that `and` and `or` return either the left or right operand based on
the truthiness of the left operand. To keep the codegen simple, we
just use C's `&&` and `||` operators, which return boolean
values (technically of type `int`, but that's not overly
important). If the operands are already booleans, there's no
inconsitency. Otherwise, we could have `42 or -5`, which should
resolve to `42`, but if we naively compile it to `42 || -5`, we get `1`
instead. Similarly, `42 and -5` should be `-5`, but in C `42 && -5`
is, again, `1`. However, `true and false` is `false` and `true &&
false` in C is also `false`.