Add block scope
Each block gets its own symbol table. For functions, this means that
the parameters go in an outer, function-level scope, while everything
else goes in an inner block scope. This allows for self-shadowing of
parameters, that is,

    func thing(x: int) -> int {
        var x := x
        x := x + 1
        return x
    }

Here, the parameter x is shadowed by a local variable set to the value
of the parameter. This is the idiomatic way to create mutable (`var`)
parameters (since paramters are immutable). Self-shadowing as a `val`
is less useful, though, since parameters are essentially `val` anyway.